feat(housing): warn when PaissaDB has no lottery data
